Update standard library docs
