+++ std/doc.go
@@ -1,0 +1,14 @@
+// Copyright 2005-2017 Marko Kungla. All rights reserved.
+// Use of this source code is governed by a Apache License 2.0
+// license that can be found in the LICENSE file.
+
+/*
+Package std is collection of libraries which extend Golang standard libraries
+and which would have name collision and therefore are all prefixed with 'h'
+to avoid need of aliasing the imports. Packages may use same same method naming
+convention, but have different behavior.
+
+*/
+package std
+
+// TODO(mkungla): Complete the docs
