+++ std/hlog/doc.go
@@ -1,4 +1,4 @@
-// Copyright 2017 Marko Kungla. All rights reserved.
+// Copyright 2005-2017 Marko Kungla. All rights reserved.
 // Use of this source code is governed by a Apache License 2.0
 // license that can be found in the LICENSE file.
