+++ std/hslices/hslices.go
