--- std/hslices/int_test.go
@@ -1,5 +1,0 @@
-// Copyright 2005-2017 Marko Kungla. All rights reserved.
-// Use of this source code is governed by a Apache License 2.0
-// license that can be found in the LICENSE file.
-
-package hslices
